Prepare API for versioning
