+++ .rubocop.yml
@@ -41,6 +41,8 @@ AllCops:
 
 Layout/LineLength:
   Max: 120
+  Exclude:
+    - lib/udatapp/api_current_container.rb
 
 Layout/MultilineMethodCallIndentation:
   EnforcedStyle: indented_relative_to_receiver
