+++ app/constants/common.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Common
+  API_CURRENT_VERSION = 'api/v1'
+
+  SIMPLE_EMAIL_REGEXP = /\A[^@\s]+@[^@\s]+\z/.freeze
+end
