+++ app/contracts/api_current/notes/create_note_contract.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module ApiCurrent
+  module Notes
+    class CreateNoteContract < BaseContract
+      json do
+        required(:title).filled(:string)
+        required(:body).filled(:string)
+        optional(:private).filled(:bool)
+      end
+    end
+  end
+end
