+++ app/contracts/api_current/notes/update_note_contract.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module ApiCurrent
+  module Notes
+    class UpdateNoteContract < BaseContract
+      json do
+        optional(:title).filled(:string)
+        optional(:body).filled(:string)
+        optional(:private).filled(:bool)
+      end
+    end
+  end
+end
