--- app/facades/notes/note_facade.rb
@@ -1,24 +1,0 @@
-# frozen_string_literal: true
-
-module Notes
-  class NoteFacade
-    def initialize(note:)
-      @note = note
-    end
-
-    delegate :id, :title, :body, to: :note
-
-    def public
-      !note.private?
-    end
-
-    def published_at
-      note.created_at.iso8601
-    end
-
-    private
-
-    # @return [Note]
-    attr_reader :note
-  end
-end
