+++ app/serializers/api_current/sessions/authentication_info_serializer.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module ApiCurrent
+  module Sessions
+    class AuthenticationInfoSerializer < BaseSerializer
+      json_schema do
+        {
+          access_token: Types::Strict::String,
+          expires_at: Types::Strict::String
+        }
+      end
+    end
+  end
+end
