--- app/services/sessions/authenticate_user.rb
@@ -1,51 +1,0 @@
-# frozen_string_literal: true
-
-module Sessions
-  class AuthenticateUser
-    include Dry::Monads[:result, :do]
-
-    include Udatapp::Import[contract: 'contracts.sessions.authenticate_user_contract']
-
-    # @param params [Hash]
-    #
-    # @return [Dry::Monads::Result<Hash, Symbol>]
-    def call(params:)
-      form = yield contract.call(params)
-
-      user = yield find_user(form[:email], form[:password])
-
-      access_token, expiration_time = encode_credentials_for(user)
-
-      serialize_result(access_token, expiration_time)
-    end
-
-    private
-
-    def find_user(email, password)
-      user = User.find_by(email: email)
-
-      return Failure(:invalid_email) unless user
-      return Failure(:invalid_password) unless user.authenticate(password)
-
-      Success(user)
-    end
-
-    def encode_credentials_for(user, expires_at: 24.hours.from_now)
-      [
-        JWT.encode(
-          { exp: expires_at.to_i, user_id: user.id },
-          Rails.application.secrets.secret_key_base
-        ),
-        expires_at
-      ]
-    end
-
-    def serialize_result(access_token, expiration_time)
-      Success(
-        Sessions::AuthenticationInfoFacade
-          .new(access_token: access_token, expiration_time: expiration_time)
-          .then { Sessions::AuthenticationInfoSerializer.new(_1).build_schema }
-      )
-    end
-  end
-end
