+++ config/initializers/import.rb
@@ -2,6 +2,8 @@
 
 require 'dry/auto_inject'
 
+require './lib/udatapp/api_current_container'
+
 module Udatapp
   Import = Dry::AutoInject(Container)
 end
